Hack.Assembler

Description

The Assembler defines the valid Hack assembly instructions and provides tools for manipulating assembly, including the actual act of assembling.

Documentation

data Instruction

A single instruction that may be executed in a single cycle of the machine

Constructors

Address Constant
A constant that will be loaded into memory

Marker String
A label that identifies a location in the program

Calculate Calculation (Maybe Destination) (Maybe Jump)
A calculation for the ALU to perform

data Constant

The constants that can be addressed by an address command

Constructors

Literal Int
A literal value

Symbol String
A symbol that will be translated into a value

data Calculation

Defines the possible calculations that can be made by the ALU

Constructors

Load0

Load1

LoadN1

LoadD

LoadA

InvD

InvA

NegD

NegA

IncD

IncA

DecD

DecA

AddDA

SubDA

SubAD

AndDA

OrDA

LoadM

InvM

NegM

IncM

DecM

AddDM

SubDM

SubMD

AndDM

OrDM

data Destination

Defines the possible locations in which to store the result of a calculation

Constructors

StoreM
Stores the result into Memory[A-register]

StoreD
Stores the result into the D-register

StoreMD

StoreA
Stores the result into the A-register

StoreAM

StoreAD

StoreAMD

data Jump

Defines the possible jumps that can occur after performing a calculation

Constructors

JumpGT
Jumps iff calc > 0

JumpEQ
Jumps iff calc == 0

JumpGE
Jumps iff calc >= 0

JumpLT
Jumps iff calc < 0

JumpNE
Jumps iff calc /= 0

JumpLE
Jumps iff calc <= 0

Jump
Jumps unconditionally

assemble :: [Instruction] -> Either [String] [Word16]

Assembles instructions to their 16-bit binary form; provides informative error strings on failure

parseAssembly :: String -> Either [String] [Instruction]

Transforms source assembly code into assembly instructions; provides informative error strings on failure