Hack.Translator

Description

The Translator provides a way to convert the specified virtual machine to assembly.

Documentation

data Command

All of the VM command types

Constructors

Add
Pops the top 2 values (STACK,x,y) from the stack and pushes their addition (x+y)

Sub
Pops the top 2 values (STACK,x,y) from the stack and pushes their subtraction (x-y)

Neg
Arithmetic negation of the top value of the stack

Eq
Pops the top 2 values (STACK,x,y) from the stack and pushes -1 if equal else 0

Gt
Pops the top 2 values (STACK,x,y) from the stack and pushes -1 if x > y else 0

Lt
Pops the top 2 values (STACK,x,y) from the stack and pushes -1 if x < y else 0

And
Pops the top 2 values (STACK,x,y) from the stack and pushes their bitwise AND

Or
Pops the top 2 values (STACK,x,y) from the stack and pushes their bitwise OR

Not
Bitwise NOT of the top value of the stack

Push Segment Int
Pushes a value from the top of the stack into the specified memory area

Pop Segment Int
Pops a value from the specified memory area to the top of the stack

Label String
Declares a function-local label

Goto String
Jumps to a label defined in the same function

IfGoto String
Jumps if the top of the stack is not 0

Function String Int
Defines a named function that has the specified number of local variables

Call String Int
Calls the named function with the specified number of arguments already on the stack

Return
Transfers flow back to the calling function

data Segment

References a certain segment of VM memory

Constructors

Argument
stores the function's arguments

Local
stores the function's local variables

Static
stores static variables shared by all functions in the same .vm file

Constant
pseudo-segment that holds al the constants in the range 0...32767

This
general-purpose segments can be made to correspond to different areas in the heap

That
serves various programming needs

Pointer
a two-entry segment that holds the base addresses of the this and that segements

Temp
fixed eight-entry segment that holds temporary variables for general use

translate :: [(String, [Command])] -> Either [String] [Instruction]

Translates VM code into Assembly; provides informative error strings on failure

parseVM :: String -> Either [String] [Command]

Transforms VM source code into its virtual representation; provides informative error strings on failure